feat: add settings screen with language picker and i18n (en/zh-CN/zh-TW)
- Add AppStrings i18n system with English, Simplified Chinese, Traditional Chinese
- Add LocaleNotifier provider with SharedPreferences persistence
- Create SettingsScreen with language dropdown + monitoring feature toggles
- Add settings entry in ConnectionScreen AppBar and LiveView top bar
- Internationalize all UI strings across connection, liveview, and assist screens

+++ android/app/build.gradle.kts
@@ -24,8 +24,8 @@ android {
         // minSdk 21 per spec: USB OTG + PTP/IP baseline.
         minSdk = flutter.minSdkVersion
         targetSdk = 34
-        versionCode = flutter.versionCode
-        versionName = flutter.versionName
+        versionCode = flutter.versionCode().toInt()
+        versionName = flutter.versionName()
     }
 
     buildTypes {
